Added bot check, if not logged in
Added functions to check if a bot (such as the Google bot) is accessing the site, and to check if the user is logged into WordPress. That way, Google doesn't see the popup and penalize your SEO. It also excludes when users are logged in, because it assumes that you have checked your users before issuing them an account.